Point 에서 사용되는 int를 CoordinateNumber 로 포장
CoordinateNumber 는 특정 범위내에서 주로 쓰인다고 가정
(해당 범위는 캐시를 적용해서 매번 새로운 인스턴스를 만들지 않도록 구현)

+++ src/main/java/coordinate/view/Board.java
@@ -1,5 +1,6 @@
 package coordinate.view;
 
+import coordinate.domain.AbstractPoint;
 import coordinate.domain.Point;
 import coordinate.domain.Points;
 
@@ -83,7 +84,7 @@ public class Board {
     public void drawPoints(Points points, char ch) {
         for (int i = 0; i < points.size(); i++) {
             Point p = points.get(i);
-            set(board, VERTICAL_START + p.getY(), HORIZON_START + (LEN_DIGITS * p.getX()), ch);
+            set(board, VERTICAL_START + p.getY().toInt(), HORIZON_START + (LEN_DIGITS * p.getX().toInt()), ch);
         }
     }
